backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
sqlalchemy==2.0.44
python-multipart==0.0.20
python-jose==3.5.0
bcrypt==5.0.0
python-decouple==3.8
pydantic-settings==2.11.0
alembic==1.17.1
passlib==1.7.4
twilio==9.8.5
openai==0.28.0
google-generativeai==0.8.5
python-dotenv==1.0.1
groq==0.11.0
httpx==0.27.0
firebase-admin
email-validator
psycopg2-binary